Fixed Bugs and Re-Implemented School of Fish
Bugs Fixed: Enemy repeat in pool, Attack index not saved, Shuffle creates variant tiles

+++ Assets/Scripts/Entity/Entity.cs
@@ -56,6 +56,6 @@ public class Entity : MonoBehaviour
 	protected virtual void OnTakeDamage(int damageTaken) { }
 	protected virtual void OnHeal(float amountHealed) { }
 
-	public float HealthPercent() => _currentHealth / _maxHealth;
+	public float HealthPercent() => _currentHealth / (float) _maxHealth;
 	public int HealthPercentReadable() => _currentHealth * 100 / _maxHealth;
 }
